+++ tools/themes/tool.js
@@ -6,7 +6,7 @@
         name:'themes',
         category: 'general',
         type:'multiple',
-        possibleValues: ['demo','justExplorer'],
+        possibleValues: ['performance','justExplorer','mobile'],
         defaultValue: '',
         advanced: true
       }
@@ -17,15 +17,17 @@
         'options.general': 'General',
         'options.themes': 'Apply a style theme',
         'themes.default':'None',
-        'themes.demo':'Demo theme',
-        'themes.justExplorer':'Just Explorer'
+        'themes.performance':'Performance theme',
+        'themes.justExplorer':'Just Explorer',
+        'themes.mobile':'Mobile theme'
       },
       'ro-RO':{
         'options.general': 'General',
         'options.themes': 'Aplic\u0103 o tem\u0103 de stil',
         'themes.default':'Nici una',
-        'themes.demo':'Tem\u0103 demo',
-        'themes.justExplorer':'Doar Explorator'
+        'themes.performance':'Tem\u0103 performan\u0163\u0103',
+        'themes.justExplorer':'Doar Explorator',
+        'themes.mobile':'Tem\u0103 mobile'
       }
     }